// src/api/generated/model/mgmtDistributionSet.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtSoftwareModule } from './mgmtSoftwareModule';
import type { Links } from './links';

/**
 * **_links**:
* **type** - The type of the distribution set
* **modules** - List of software modules
* **metadata** - List of metadata

 */
export interface MgmtDistributionSet {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  /** The technical identifier of the entity */
  id: number;
  /** Package version */
  version?: string;
  /** The type of the distribution set */
  type?: string;
  /** The type name of the distribution set */
  typeName?: string;
  /** True of the distribution set software module setup is complete as defined by the
distribution set type */
  complete?: boolean;
  /** If the distribution set is locked */
  locked?: boolean;
  /** Deleted flag, used for soft deleted entities */
  deleted?: boolean;
  /** True by default and false after the distribution set is invalidated by the user */
  valid?: boolean;
  /** True if DS is a required migration step for another DS. As a result the DS’s assignment will not be cancelled
when another DS is assigned (note: updatable only if DS is not yet assigned to a target) */
  requiredMigrationStep?: boolean;
  modules?: MgmtSoftwareModule[];
  _links?: Links;
}