// src/api/generated/model/mgmtAction.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtActionForceType } from './mgmtActionForceType';
import type { Links } from './links';

export interface MgmtAction {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  /** ID of the action */
  id?: number;
  /** Type of action */
  type?: string;
  /** Status of action */
  status?: string;
  /** Detailed status of action */
  detailStatus?: string;
  forceTime?: number;
  forceType?: MgmtActionForceType;
  /** Weight of the action showing the importance of the update */
  weight?: number;
  /** The ID of the rollout this action was created for */
  rollout?: number;
  /** The name of the rollout this action was created for */
  rolloutName?: string;
  /** (Optional) Code provided as part of the last status update that was sent by the device. */
  lastStatusCode?: number;
  /** If created by external system this field contains the external reference for the action */
  externalRef?: string;
  _links?: Links;
}